pyproject.toml
[project]
name = "de-coded-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.3.1",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "astunparse==1.6.3",
    "bcrypt==4.1.3",
    "black==25.9.0",
    "boto3==1.40.50",
    "botocore==1.40.50",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cryptography==46.0.2",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.110.1",
    "flake8==7.3.0",
    "flatbuffers==25.9.23",
    "gast==0.6.0",
    "google-pasta==0.2.0",
    "grpcio==1.75.1",
    "h11==0.16.0",
    "h5py==3.15.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "imagehash==4.3.2",
    "imageio==2.37.0",
    "iniconfig==2.1.0",
    "isort==6.1.0",
    "jiter==0.11.0",
    "jmespath==1.0.1",
    "jq==1.10.0",
    "keras==3.11.3",
    "keras-applications==1.0.8",
    "lazy-loader==0.4",
    "libclang==18.1.1",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.3",
    "motor==3.3.1",
    "mypy==1.18.2",
    "mypy-extensions==1.1.0",
    "namex==0.1.0",
    "networkx==3.5",
    "numpy==2.2.6",
    "oauthlib==3.3.1",
    "openai==2.3.0",
    "opencv-python-headless==4.12.0.88",
    "opt-einsum==3.4.0",
    "optree==0.17.0",
    "packaging==24.2",
    "pandas==2.3.3",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "pillow==12.0.0",
    "pinecone==7.3.0",
    "pinecone-plugin-assistant==1.8.0",
    "pinecone-plugin-interface==0.0.7",
    "platformdirs==4.5.0",
    "pluggy==1.6.0",
    "protobuf==6.33.0",
    "pyasn1==0.6.1",
    "pycodestyle==2.14.0",
    "pycparser==2.23",
    "pydantic==2.12.0",
    "pydantic-core==2.41.1",
    "pyflakes==3.4.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pymongo==4.5.0",
    "pytest==8.4.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pytokens==0.1.10",
    "pytz==2025.2",
    "pywavelets==1.9.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rich==14.2.0",
    "rsa==4.9.1",
    "s3transfer==0.14.0",
    "s5cmd==0.2.0",
    "scikit-image==0.25.2",
    "scipy==1.16.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.37.2",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.20.0",
    "termcolor==3.1.0",
    "tifffile==2025.10.4",
    "tqdm==4.67.1",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.25.0",
    "watchfiles==1.1.0",
    "werkzeug==3.1.3",
    "wrapt==1.17.3",
]

[project.scripts]
start = "uvicorn main:app --port 8000"
dev = "uvicorn main:app --reload --port 8000"